Move timing to wpimath DevMain.java
